requirements.txt
appdirs==1.4.4
atomicwrites==1.4.0
attrs==20.3.0
beautifulsoup4==4.9.3
certifi==2020.12.5
cfgv==3.2.0
chardet==4.0.0
colorama==0.4.4
coverage==5.3.1
distlib==0.3.1
dnspython==2.1.0
filelock==3.0.12
frozendict==1.2
identify==1.5.13
idna==2.10
iniconfig==1.1.1
marshmallow==3.10.0
mypy==0.790
mypy-extensions==0.4.3
nodeenv==1.5.0
packaging==20.8
pluggy==0.13.1
pre-commit==2.10.0
psycopg2==2.8.6
py==1.10.0
pymongo==3.11.2
pyparsing==2.4.7
pytest==6.2.1
pytest-cov==2.10.1
python-dateutil==2.8.1
python-dotenv==0.15.0
PyYAML==5.4.1
requests==2.25.1
selenium==3.141.0
six==1.15.0
soupsieve==2.1
toml==0.10.2
typed-ast==1.4.2
typing-extensions==3.7.4.3
urllib3==1.26.2
virtualenv==20.4.0
